Alertmanager: Add OpsGenie Webhook notification.
This adds notifications to OpsGenie. It is using the webhook api. The
notification_config should look like

  opsgenie_config {
    api_key: "bliep"
    teams: "team-one"
    send_resolved: true
  }

It further sets the alert fingerprint in Alias and adds all the
alert labels into the Details field.

See https://www.opsgenie.com/docs/web-api/alert-api#createAlertRequest
for the details.

It can also close the alert by using the Alias field again. See
https://www.opsgenie.com/docs/web-api/alert-api#closeAlertRequest

And it adds a 'labels_to_tag' that allows you to specify for which
label names you want to create a tag with the value of that label.
This makes routing alerts easier.