Alertmanager: Add Amazon SNS Notification
This adds notifications to Amazon SNS.  It is using Amazon's go sdk. The notification config should look like
    amazon_sns_config {
      topic_arn: "arn:aws:sns:us-east-1:an-arn-id:topic"
      send_resolved: true
    }

The alert summary and status will be used as the SNS subject.
The alert description and status will be used as the SNS message.

SNS notifictions require Amazon AWS credentials to be setup as described by the Amazon go sdk.