Add postgres storage backend and allow selecting providers.
This commit does an initial refactor of the code to support the ql,mem
and a new postgres backend as backing storage for the alertmanager.

The postgres backend is introduced to provide a persistent storage layer that
avoids QL, which may have concurrency bugs at the present time.

The storage backend selection code is prelimininary, and probably needs some
architecting to make it a more modular for the frontend.