Add route tests for receiver inheritance
Tests for:
- inheriting a receiver from the default route
- overriding the receiver when the parent route has not set one
- inheriting a receiver from a parent route that's not the default

Adds a route which matches on a `group_by` label which inherits the
default receiver and also has a child route, matching on an `env` label,
which overrides the receiver.

To be sure of edge case behaviour, we test inheriting the receiver from
a child route. Hopefully no one would actually implement this edge case
in reality as it's overly complex but we test anyway to be sure that it
doesn't trigger unexpected behaviour.