Cease using lastID and send totalSilences
Last ID was broken for jumping between pages.
Using the offset and number of requested silences
is enough. The number of silences existing is now
upgraded and dynamically reflected in the UI, as
well.

Unimportant implementation detail: Now using int
instead of uint64 for offset+n, since the original
reason for doing this (consistency with UUID's
being uint64) is no longer the case.