Simplify inconsistent silence update behavior
The old silence update behavior resulted in duplicated silences after
e.g. updating a typo in a silence comment. I have to admit that I don't
understand the idea behind the old logic, but I couldn't see a use case
for creating a second silence when clicking on "Edit". The new behavior
is very simple and will simply always delete an existing silence and
replace it with the updated behavior.

Fixes #188.