Include receiver name in notification metrics
Add the receiver name to the `alertmanager_notifications_failed_total`
and `alertmanager_notifications_total` metrics so that it's possible to
identify which receiver is causing errors.

For example, I might have two webhook receivers, one for sending
notifications to [Jiralerts][] and another to send alerts to
Elasticsearch for reporting purposes. If the `webhook` integration
starts alerting, the receiver name allows you to determine which service
is causing the errors.

[Jiralerts]: https://github.com/fabxc/jiralerts

Example receiver configuration:

    receivers:
    - name: 'my_webhook'
      webhook_configs:
      - url: 'http://127.0.0.1:5001/'

Example metrics produced:

    alertmanager_notifications_failed_total{integration="webhook",receiver="my_webhook"} 2
    alertmanager_notifications_total{integration="webhook",receiver="my_webhook"} 1