Send the prom gen url as alert source
From the documentation at
https://v2.developer.pagerduty.com/docs/send-an-event-events-api-v2

payload.source: required string

The unique location of the affected system, preferably a hostname or FQDN.

Sending the generatorURL isn't technically the
affected system, but an event sent to PD can
encompass multiple alerts. The generatorURL seems
like the correct way to help users debug this.