#1196 Implement ability to get createdBy from header
If a header is provided (most likely by a load balancer) that states who
the user is then we should use that value for the createdBy field for
proper accounting/auditability of silence creation.

The header used can be configured by setting the `auth_header` option in
the global context. The feature can be disabled by setting `auth_header:
""`.

The value set for `auth_header` must match the output from the go http
request parser

```
HTTP defines that header names are case-insensitive. The
request parser implements this by using CanonicalHeaderKey,
making the first character and any characters following a
hyphen uppercase and the rest lowercase.
```