Remove Store interface
The interface was only used once, and defined
within the actual store package. Besides not being
necessary (and being very simple to define), the
current best practice is to define the interface
by the calling code (what the caller requires)
rather than the implemented code (in this case,
the store package).

Signed-off-by: stuart nelson <stuartnelson3@gmail.com>