feat(api): add /api/version handler for getting version
With API v2 there are now 2 API endpoints a client could use, but there is no dicovery mechanism to tell which API version are avaiable, so a client needs to make assumptions or try all possible versions.
Part of the problem is that there is no non-versioned API endpoint that exposes version information, it's only avaiable via api/v*/status.
This adds api/version endpoint that's only returing alertmanager version, this information can be used by the client to decide which API version to use. It doesn't fully resolve discoverability problem since it requires a client to keep information about all Alertmanager releases, but it helps while being useful on it's own.

Signed-off-by: Łukasz Mierzwa <l.mierzwa@gmail.com>