Use setAlertStatusFn
This encapsulates the actual logic being done
previously into a single closure that gets passed
to the api when loaded/reloaded.

Additionally, the naming is now a bit clearer. The
use of the Muter interface is a bit confusing as
it was previously being used for its side effect
of marking an alert as inhibited or not, and the
return value was being ignored.

Signed-off-by: stuart nelson <stuartnelson3@gmail.com>