WIP: Silence based on values outside the alert labels
Sometimes it's desirable to silence the alerts that come from
a specific source, e.g. a single Prometheus instance that does
behave in a funny way and turning it off or using netfilter is
not an option.

Expand the the Muter interface to consume the alert labels and
an extra set of labels. Use an extra argument to prevent a copy
the model.LabelSet. Use a flat namespace but prefix the extrat
symbols using __. Turn the generatorURL into an extra label and
update the notify test.

Related: #344