config/notifiers: PagerDuty details override defaults
According to the Alertmanager documentation, the `details` option of
`pagerduty_config` defaults to

```
[ details: { <string>: <tmpl_string>, ... } | default = {
  firing:       '{{ template "pagerduty.default.instances" .Alerts.Firing }}'
  resolved:     '{{ template "pagerduty.default.instances" .Alerts.Resolved }}'
  num_firing:   '{{ .Alerts.Firing | len }}'
  num_resolved: '{{ .Alerts.Resolved | len }}'
} ]
```

Which suggests that if `details` is provided in the configuration, the entire
option is overridden.

But instead, the individual keys of `details` are being set to the values above
if the corresponding key isn't present in the configuration. This means it's
impossible to remove those keys entirely; setting them to an empty string would
still have them displayed in PagerDuty.

This commit implements the documented behavior, i.e. it sets
`PagerDutyConfig.Details` to `DefaultPagerdutyDetails` only if the former is
`nil`.

Signed-off-by: Benoît Knecht <bknecht@protonmail.ch>