feat: add silence logging option
This code was originally written by m-masataka, but tweaked by
anarcat, and re-tweaked by @coderabbitai, which seems to be an
automated AI test suite that was introduced on the project recently.

Tests were written by TheMeier.

I reviewed the AI output and could not find any better way to do
things.

Signed-off-by: m-masataka <m.mizukoshi.wakuwaku@gmail.com>
Signed-off-by: Antoine Beaupré <anarcat@debian.org>
Signed-off-by: Solomon Jacobs <solomonjacobs@protonmail.com>