Alert store not calling callback when an alert is deleted
The garbage collection process within the store is in charge of
determining if an alert is resolved, deleting it, and then communicating
this back to the callback set.

When an alert was explicitly deleted, these were not being communicated
back to the callback and caused the metric to report incorrect results.

Fixes #2619

Signed-off-by: gotjosh <josue@grafana.com>