amtool: Detect version drift and warn users (#2672)
* amtool: Detect version drift and warn users

This change detects the alertmanager version when initiating the client.
It ignores most errors since I expect amtool to fail later.

If amtool is not compiled with proper version, we do not do anything
either.

We use MajorMinor for now as we have not reach 1.0, but we still allow
the bugfix version number (Z in x.y.Z) to differ.

Signed-off-by: Julien Pivotto <roidelapluie@inuits.eu>

* Add version check

Signed-off-by: Julien Pivotto <roidelapluie@inuits.eu>
Signed-off-by: Kyle Brandt <kyle@grafana.com>