Add support for custom Go template delimiters
Alertmanager uses Go templates in different configuration parameters.
Many users generate the configuration using tools like Helm, ExternalSecret, etc.
Most of these tools are written in Go, and therfore use Go templates.
This creates problems when the configuration tool needs to skip the Alertmanager templates.
As a result users usually have to add escaping to the templates:
```go
{{ `{{ ... }}` }}
```

This change allows setting the template delimiters used in Alertmanager configuration.
This is already configurable in Go templates using `Delims()` functions for both `text` and `html`.
The newly added global configuration parameters allows users to set the left and right delimiters
used in Alertmanager templates and therfore avoid conflict with other configuration management tools:

```yaml
template_config:
  delim_left: "[["
  delim_right: "]]"
```

This change is backward compatible as empty delimiter parameters stands for the corresponding default: `{{` or `}}`.
The builtin templates are not affected by this change as custom delimiters are only applied to:
- `<tmpl_string>`
- `<tmpl_secret>`
- custom templates loaded by `templates: []`

Signed-off-by: Siavash Safi <siavash.safi@gmail.com>