Capture receiver integration errors
As an alerts traverses through the routing tree, there's a chance we fail to delivery the notification as we attempt to send it.

This creates a problem where a user that has a failure in their template or their receiver integration (e.g. webhook) is unavailable has no visibility on this problem.

This change, enhances the receiver API to include the last error, duration and time an alert went through the receiver.

Signed-off-by: gotjosh <josue.abreu@gmail.com>