Handle errors in Slack's postMessage API
For larger organisations, using the postMessage API has advantages over
webhooks, such as not being tied to a specific user (who might leave the
company at some point, resulting in the webhook integration becoming
disabled), and posting to multiple channels.

Alertmanager's Slack integration actually works with this endpoint out of the
box, by setting the URL path to /api/chat.postMessage. However, error
handling works a bit differently on this endpoint, always returning 200 OK
even when the posted message is rejected.

This commit handles the response format from this endpoint, so that errors are
reported correctly.