Gate receiver names behind a feature flag
Alertmanager doesn't actually have a feature flag system yet, so this
invents the universe, mostly by cribbing off the Prometheus model.

Here we introduce a new CLI flag `--enable-feature` that we parse into
a new `featureConfigs` struct. We also update the PipelineBuilder
to take that flag and pass it through the metrics infrastructure. This
also required a bit of deduplication to make finding the correct
labelset (with or without the receiver name) easier in every place that
it's used, and updating the tests to use that same mechanism.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>