Improve the notify error log.
Improve the notify error log by including the number of alerts that the
notification attempt included, as well as a summary of the names of
the alerts. In most cases, the number of alerts being notified on is
small, so the summary will contain all the alert names. Including the
alert names can give the operator a hint on where to look to diagnose
a notification problem.

Signed-off-by: James Peach <jpeach@cloudflare.com>