Fix unexpected behavior interpreting matchers in amtool
This commit fixes some of the unexpected behavior when interpreting
matchers in amtool.

Before this commit, <matcher-groups> that looked like a valid matcher
such as {foo=bar} would have "alertname" prefixed as amtool considered
it to be a label value. However, this would create a silence for
alertname="{foo=bar}" which is almost never want the user intended.

Instead, this commit updates the heuristic to be more accurate in
determining whether the input is a label value without a label name
and operator or if it is a complete matcher that should be parsed
from start to end.

Signed-off-by: George Robinson <george.robinson@grafana.com>